requirements.txt
Pillow
mcschematic
numpy
future
ffmpeg-python